Refactoring: Extracted code from one_dim_chainer_tool into new class.
So that we cann call that code from our new condense compress tool.
